refactoring v1.0.0

+++ app/src/main/java/com/victoria/bleled/app/test/animation/DialView.kt
@@ -91,7 +91,7 @@ class DialView @JvmOverloads constructor(
             FanSpeed.LOW -> fanSpeedLowColor
             FanSpeed.MEDIUM -> fanSpeedMediumColor
             FanSpeed.HIGH -> fanSpeedMaxColor
-        } as Int
+        }
         canvas?.drawCircle((width / 2).toFloat(), (height / 2).toFloat(), radius, paint)
 
         // draw the indicator circle
